+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ CHANGELOG
 
 - Added attribute cookies to class Request()
   [masipcat]
+- Added uvicorn as a guillotina requirement
+  [masipcat]
 
 
 6.0.0a1 (2019-12-09)