+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 5.1.23 (unreleased)
 -------------------
 
+- Make pytest.mark.app_settings work in older pytest versions too [lferran]
+
 - @move: destination id conflict should return 409 error, not 412
   [inaki]